Average Precipitations
This program was intended to take the input of a file on rainfall by
month and average them.  THIS PROGRAM DOES NOT RUN CORRECTLY.  The
output is only for one year, it should be for as many years as you
want.  This is because the month is read as a string, not an integer.